+++ crates/type-system/src/ontology/data_type/error.rs
@@ -11,7 +11,7 @@ use serde::{Deserialize, Serialize};
 #[cfg(target_arch = "wasm32")]
 use tsify::Tsify;
 
-use crate::{ontology::shared::HasSerdeJsonError, uri::ParseVersionedUriError};
+use crate::ontology::{uri::ParseVersionedUriError, HasSerdeJsonError};
 
 #[cfg_attr(target_arch = "wasm32", derive(Tsify))]
 #[derive(Debug, Deserialize, Serialize)]
