+++ crates/type-system/src/ontology/shared/mod.rs
@@ -1,20 +1,5 @@
-mod array;
-mod error;
-mod object;
-mod one_of;
-mod validate;
-
-pub use array::*;
-pub use error::*;
-pub use object::*;
-pub use one_of::*;
-pub use validate::*;
-
-// Re-export the repr contents so they're nicely grouped and imports are less confusing
-pub mod repr {
-    pub use super::{
-        array::repr::{Array, ValueOrArray},
-        object::repr::Object,
-        one_of::repr::OneOf,
-    };
-}
+pub mod array;
+pub mod error;
+pub mod object;
+pub mod one_of;
+pub mod validate;
