+++ crates/type-system/src/ontology/shared/one_of/mod.rs
@@ -1,4 +1,4 @@
-pub(super) mod repr;
+pub(in crate::ontology) mod repr;
 
 use crate::ontology::shared::validate::ValidationError;
