+++ crates/type-system/src/ontology/shared/validate.rs
@@ -8,7 +8,7 @@ use serde::{Deserialize, Serialize};
 #[cfg(target_arch = "wasm32")]
 use tsify::Tsify;
 
-use crate::uri::{BaseUri, VersionedUri};
+use crate::ontology::uri::{BaseUri, VersionedUri};
 
 #[cfg_attr(target_arch = "wasm32", derive(Tsify))]
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
